+++ src/views/components/gcp-config-card.ts
@@ -1,5 +1,3 @@
-import { ADC_FILE_PATH } from "../../constants";
-import { configNameToTitle, readJsonFile } from "../../helpers";
 import { GCP_CONFIGURATION } from "../../types";
 
 type GcpConfigCardProps = {
@@ -16,11 +14,21 @@ export const gcpConfigCard = ({
           <input 
             class="gcp-config-badge" 
             type="radio" name="gcpConfig" 
-            onclick="handleSwitchProjectClick(this);" 
+            onclick="handleSwitchConfigClick(this);" 
             value="${gcpConfigIndex}" 
             ${gcpConfig.is_active ? "checked" : ""}
           />
-          <h3 class="gcp-config-title">${configNameToTitle(gcpConfig.name)}</h3>
+          <h3 class="gcp-config-title">
+            ${gcpConfig.name}
+            <div class="gcp-config-edit-action">
+              <button class="button-icon button-text"  onclick="handleEditConfigClick(${gcpConfigIndex});">
+                <i class='codicon codicon-edit'></i>
+              </button>
+              <button class="button-icon button-text"  onclick="handleDeleteConfigClick(${gcpConfigIndex});">
+                <i class='codicon codicon-trash'></i>
+              </button>
+            </div>
+          </h3>
           <div class="gcp-config-info">
               <div class="gcp-config-info-item">
                   <small class="title">Project Id</small>